# Conflicts:
#	DoAnTotNghiepBE.iml
#	target/classes/application.properties

+++ src/main/java/com/example/doantotnghiepbe/repository/PriceRepository.java
@@ -3,6 +3,6 @@ package com.example.doantotnghiepbe.repository;
 import com.example.doantotnghiepbe.entity.Price;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PriceRepository extends JpaRepository<Price, Integer> {
+public interface PriceRepository extends JpaRepository<Price, Long> {
     Price findByPriceId(Integer priceId);
 }
